% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allOnOne.R
\name{ekbohmMLE}
\alias{ekbohmMLE}
\title{Ekbohm MLE based t test}
\usage{
ekbohmMLE(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{numerical vector, should be same length as y}

\item{y}{numerical vector, should be same length as x}

\item{alternative}{'Greater', 'Less', or 'Two.Sided' based on direction of your test}
}
\description{
Ekbohm MLE based t test
}
