% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allOnOne.R
\name{linstiversMLE}
\alias{linstiversMLE}
\title{Lin And Stivers MLE based t test}
\usage{
linstiversMLE(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{numerical vector, should be same length as y}

\item{y}{numerical vector, should be same length as x}

\item{alternative}{'Greater', 'Less', or 'Two.Sided', based on the direction of your test}
}
\value{
returns warnings if input criteria is not met, or if results
may be ineffective, returns test statistic and p value if all is correct
}
\description{
Lin And Stivers MLE based t test
}
\examples{
x <- c(6,9,12,15,NA,13,17,NA,13,18,NA) y <- c(10,NA,16,15,14,12,12,17,NA,11,12)
linstiversMLE(x,y,alternative = 'two.sided')
}
