% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allOnOne.R
\name{modifiedTstatistic}
\alias{modifiedTstatistic}
\title{Kim et al Modified T-statistic}
\usage{
modifiedTstatistic(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{numerical vector, should be same length as y}

\item{y}{numerical vector, should be same length as x}

\item{alternative}{'Greater', 'Less', or 'Two.Sided' based on direction of your test}
}
\value{
returns warnings if input criteria is not met, or if results
may be ineffective, returns test statistic and p value if all is correct
}
\description{
Kim et al Modified T-statistic
}
